Decompile/document strings

+++ payload/include/graphics.h
@@ -2,24 +2,24 @@
 #define GUARD_GRAPHICS_H
 
 extern const u32 gTitle_Gfx[];
-extern const u16 gUnknown_0201A450[];
+extern const u32 gUnknown_0201A450[];
 extern const u32 gTitle_Tilemap[];
-extern const u16 gUnknown_0201A894[];
+extern const u32 gUnknown_0201A894[];
 extern const u32 gUnknown_0201A8B4[];
 extern const u32 gUnknown_0201AD0C[];
 extern const u32 gUnknown_0201B0FC[];
-extern const u16 gUnknown_0201B1A0[];
+extern const u32 gUnknown_0201B1A0[];
 extern const u32 gUnknown_0201B2F8[];
 extern const u32 gUnknown_0201B430[];
 extern const u32 gSummaryScreen_Gfx[];
 extern const u32 gUnknown_0201BFF8[];
-extern const u16 gUnknown_0201C10C[];
+extern const u32 gUnknown_0201C10C[];
 extern const u32 gSummaryScreen_Skills_Tilemap[];
 extern const u32 gSummaryScreen_Moves_Tilemap[];
 extern const u32 gTypeStatusIcons_Gfx[];
-extern const u16 gTypeStatusIcons_Pal[];
+extern const u32 gTypeStatusIcons_Pal[];
 extern const u32 gBattleScreen_Gfx[];
-extern const u16 gBattleScreen_Pal[];
+extern const u32 gBattleScreen_Pal[];
 extern const u32 gBattleScreen_Tilemap[];
 extern const u32 gBattleWindows_Tilemap[];
